+++ frontend/app/[locale]/components/layout/baseLayout.tsx
@@ -4,12 +4,14 @@ import Footer from '../footers/footer'
 
 const BaseLayout = ({ children }: {children: React.ReactNode}) => {
     return (
-        <div className='min-h-screen flex flex-col
+        <div className='min-h-screen flex flex-col justify-between
             bg-gradient-to-tr from-fuchsia-50 via-purple-300 to-cyan-300'>
             <Headers />
-            <div className='mt-16'>
+
+            <main className={`mb-auto py-16`}>
                 {children}
-            </div>
+            </main>
+
             <Footer />
         </div>
     )